+++ social/graphql/mutation.py
@@ -1,5 +1,4 @@
 from django.db import models
-from django.contrib.auth.models import User
 import datetime
 import graphene
 from graphene_django import DjangoObjectType
