+++ user/graphql/mutation.py
@@ -9,37 +9,8 @@ from interests.graphql.types import InterestType
 from social.graphql.types import FriendType, FriendshipRequestType
 from system.graphql.types import ImageType
 from system.graphql.mutation import create_system_image
-from .types import InviteUserType, UserType, ProfileType, AuthUserType
-from ..models import Profile, InviteUser
-
-
-class CreateInvite(graphene.Mutation):
-    invite_user = graphene.Field(InviteUserType)
-    sponsor = graphene.String()
-
-    class Arguments:
-        name = graphene.String(required=True)
-        email = graphene.String(required=False)
-        # phone = graphene.String(required=False)
-        note = graphene.String(required=False)
-        avatar = graphene.String(required=False)
-
-    def mutate(self, info, name, email, note, avatar):
-        sponsor = info.context.user.username
-        invite_user = InviteUser.objects.create(
-            name=name,
-            email=email,
-            # phone_number=phone,
-            note=note,
-            avatar=avatar,
-            sponsor=sponsor,
-        )
-        invite_user.save()
-        return CreateInvite(invite_user=invite_user, sponsor=sponsor)
-
-
-class CreateInviteMutation(graphene.ObjectType):
-    create_invite = CreateInvite.Field()
+from .types import UserType, ProfileType, AuthUserType
+from ..models import Profile
 
 
 class LoginUser(graphene.Mutation):
